[feat] 라운드 종료 시 나오는 스코어 테이블 추가
