+++ src/components/score-table/score-table.html
@@ -40,7 +40,7 @@
           <thead>
             <tr>
               <th colspan="3" class="text-center text-lg pt-5">
-                라운드 별 우승자
+                라운드 별 승자
               </th>
             </tr>
             <tr>
