+++ src/components/score-table/test-ingame.html
@@ -1,3 +1,4 @@
+<!-- TODO 추후 파일 삭제 예정 -->
 <!doctype html>
 <html lang="ko-KR">
   <head>
@@ -12,7 +13,7 @@
     <script type="module" src="/src/pages/ingame/chat.ts"></script>
     <script
       type="module"
-      src="/src/components/score-table/test-table.ts"
+      src="/src/components/score-table/score-table.ts"
     ></script>
   </head>
   <body
@@ -51,7 +52,7 @@
             <thead>
               <tr>
                 <th colspan="3" class="text-center text-lg pt-5">
-                  라운드 별 우승자
+                  라운드 별 승자
                 </th>
               </tr>
               <tr>
